[ADD] LetterRepository 생성(#490)
