--- Manito/Manito/Data/Network/API/LetterAPI.swift
@@ -1,38 +1,0 @@
-//
-//  LetterAPI.swift
-//  Manito
-//
-//  Created by SHIN YOON AH on 2022/08/31.
-//
-
-import Foundation
-
-struct LetterAPI: LetterProtocol {
-    private let apiService: Requestable
-
-    init(apiService: Requestable) {
-        self.apiService = apiService
-    }
-    
-    func fetchSendLetter(roomId: String) async throws -> Letter? {
-        let request = LetterEndPoint
-            .fetchSendLetter(roomId: roomId)
-            .createRequest()
-        return try await self.apiService.request(request)
-    }
-    
-    func fetchReceiveLetter(roomId: String) async throws -> Letter? {
-        let request = LetterEndPoint
-            .fetchReceiveLetter(roomId: roomId)
-            .createRequest()
-        return try await self.apiService.request(request)
-    }
-    
-    @discardableResult
-    func dispatchLetter(roomId: String, image: Data? = nil, letter: LetterDTO, missionId: String) async throws -> Int {
-        let request = LetterEndPoint
-            .dispatchLetter(roomId: roomId, image: image, letter: letter, missionId: missionId)
-            .createRequest()
-        return try await self.apiService.request(request)
-    }
-}
